requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
httpx==0.28.1
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml_html_clean==0.4.1
pypdf==5.3.1
gTTS==2.5.4
edge-tts>=7,<8
numpy==1.26.4
lameenc==1.8.1
torch>=2.2,<3
omegaconf==2.3.0
